Standalone branch

+++ vstgui/lib/ccolor.cpp
@@ -1,36 +1,6 @@
-//-----------------------------------------------------------------------------
-// VST Plug-Ins SDK
-// VSTGUI: Graphical User Interface Framework for VST plugins
-//
-// Version 4.3
-//
-//-----------------------------------------------------------------------------
-// VSTGUI LICENSE
-// (c) 2015, Steinberg Media Technologies, All Rights Reserved
-//-----------------------------------------------------------------------------
-// Redistribution and use in source and binary forms, with or without modification,
-// are permitted provided that the following conditions are met:
-// 
-//   * Redistributions of source code must retain the above copyright notice, 
-//     this list of conditions and the following disclaimer.
-//   * Redistributions in binary form must reproduce the above copyright notice,
-//     this list of conditions and the following disclaimer in the documentation 
-//     and/or other materials provided with the distribution.
-//   * Neither the name of the Steinberg Media Technologies nor the names of its
-//     contributors may be used to endorse or promote products derived from this 
-//     software without specific prior written permission.
-// 
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-// ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
-// WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
-// IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
-// INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
-// BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, 
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF 
-// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE 
-// OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE  OF THIS SOFTWARE, EVEN IF ADVISED
-// OF THE POSSIBILITY OF SUCH DAMAGE.
-//-----------------------------------------------------------------------------
+// This file is part of VSTGUI. It is subject to the license terms 
+// in the LICENSE file found in the top-level directory of this
+// distribution and at http://github.com/steinbergmedia/vstgui/LICENSE
 
 #include "ccolor.h"
 #include <cmath>
@@ -161,9 +131,9 @@ void CColor::fromHSL (double hue, double saturation, double lightness)
 	r = (r + m) * 255;
 	g = (g + m ) * 255.;
 	b = (b + m ) * 255.;
-	red = (uint8_t)floor (r + 0.5);
-	green = (uint8_t)floor (g + 0.5);
-	blue = (uint8_t)floor (b + 0.5);
+	red = static_cast<uint8_t> (floor (r + 0.5));
+	green = static_cast<uint8_t> (floor (g + 0.5));
+	blue = static_cast<uint8_t> (floor (b + 0.5));
 }
 
 //-----------------------------------------------------------------------------
@@ -178,9 +148,9 @@ void CColor::toHSV (double& hue, double& saturation, double& value) const
 	}
 
 	/* Normalize value to 1 */
-	double r = (double)(red / 255.) / value;
-	double g = (double)(green / 255.) / value;
-	double b = (double)(blue / 255.) / value;
+	double r = (red / 255.) / value;
+	double g = (green / 255.) / value;
+	double b = (blue / 255.) / value;
 	double rgbMin = min3<double> (r, g, b);
 	rgbMax = max3<double> (r, g, b);
 
@@ -228,7 +198,7 @@ void CColor::fromHSV (double hue, double saturation, double value)
 		value = 1.;
 	if (saturation <= 0.)
 	{
-		red = green = blue = (uint8_t) (value * 255.);
+		red = green = blue = static_cast<uint8_t> (value * 255.);
 		return;
 	}
 	else if (saturation > 1.)
@@ -239,7 +209,7 @@ void CColor::fromHSV (double hue, double saturation, double value)
 		hue += 360.;
 
 	const double hf  = hue / 60.0;
-	const int32_t i  = (int32_t) floor ( hf );
+	const int32_t i  = static_cast<int32_t> (floor (hf));
 	const double f   = hf - i;
 	const double pv  = value * ( 1 - saturation );
 	const double qv  = value * ( 1 - saturation * f );
@@ -311,9 +281,9 @@ void CColor::fromHSV (double hue, double saturation, double value)
 			break;
 		}
 	}
-	red = (uint8_t) floor (r * 255. + 0.5);
-	green = (uint8_t) floor (g * 255. + 0.5);
-	blue = (uint8_t) floor (b * 255. + 0.5);
+	red = static_cast<uint8_t> (floor (r * 255. + 0.5));
+	green = static_cast<uint8_t> (floor (g * 255. + 0.5));
+	blue = static_cast<uint8_t> (floor (b * 255. + 0.5));
 }
 
 } // namespace
